+++ README.md
@@ -1,2 +1,2 @@
-# cryptobox-java
+# cryptology-java
 libsodium compatible SealedBox and SecretBox implementation in Java
