+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'cryptobox'
+rootProject.name = 'cryptology'
