+++ src/main/java/io/xconn/cryptology/HSalsa20.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.xconn.cryptobox;
+package io.xconn.cryptology;
 
 import java.nio.charset.StandardCharsets;
