+++ src/main/java/io/xconn/cryptology/Util.java
@@ -1,4 +1,4 @@
-package io.xconn.cryptobox;
+package io.xconn.cryptology;
 
 import java.security.SecureRandom;
